feat(withLatestFrom): add higher-order lettable version of withLatestFrom
