+++ src/operators/index.ts
@@ -78,4 +78,5 @@ export { windowCount } from './windowCount';
 export { windowTime } from './windowTime';
 export { windowToggle } from './windowToggle';
 export { windowWhen } from './windowWhen';
+export { withLatestFrom } from './withLatestFrom';
 export { zip } from './zip';
